+++ src/constants.ts
@@ -1,3 +1,4 @@
 export const EXTENSION_ID = "gistpad";
 export const FS_SCHEME = "gist";
+export const UNTITLED_SCHEME = "untitled";
 export const ZERO_WIDTH_SPACE = "‎‎\u200B"